Draft refactor
