+++ test/surface/concurrency.test.ts
@@ -8,27 +8,26 @@ import { getAuthFetcher } from "solid-auth-fetcher";
 import {
   recursiveDelete,
   getContainerMembers,
-  WPSClient,
   responseCodeGroup,
   ifWps,
+  getStore,
 } from "../helpers/util";
-import { getStore } from "../helpers/util";
+import { WPSClient } from "../helpers/NotificationsClient";
 import { ldp, rdf, space, link } from "rdf-namespaces";
 import * as rdflib from "rdflib";
 
 const waittime = 2000;
 
 if (process.env.SKIP_CONC) {
-	console.log("Skipping concurrency tests because SKIP_CONC env var is set");
-    describe('Check for SKIP_CONC env', () => {
-        test('SKIP_CONC env var is set', () => {
-            expect(true).toBe(true);
-        });
+  console.log("Skipping concurrency tests because SKIP_CONC env var is set");
+  describe("Check for SKIP_CONC env", () => {
+    test("SKIP_CONC env var is set", () => {
+      expect(true).toBe(true);
     });
+  });
 } else {
-
-// when the tests start, exists/exists.ttl exists in the test folder,
-// and nothing else.
+  // when the tests start, exists/exists.ttl exists in the test folder,
+  // and nothing else.
 
   describe("Concurrency", () => {
     let authFetcher;
@@ -38,7 +37,7 @@ if (process.env.SKIP_CONC) {
 
     // use `${testFolderUrl}exists/` as the existing folder:
     describe("Try to create the same resource, using PUT 10 times", () => {
-      const {testFolderUrl} = generateTestFolder();
+      const { testFolderUrl } = generateTestFolder();
       let websocketsPubsubClientContainer;
       let websocketsPubsubClientResource;
       let results;
@@ -58,11 +57,14 @@ if (process.env.SKIP_CONC) {
         });
 
         websocketsPubsubClientContainer = new WPSClient(
-            containerUrl,
-            authFetcher
+          containerUrl,
+          authFetcher
         );
         await websocketsPubsubClientContainer.getReady();
-        websocketsPubsubClientResource = new WPSClient(resourceUrl, authFetcher);
+        websocketsPubsubClientResource = new WPSClient(
+          resourceUrl,
+          authFetcher
+        );
         await websocketsPubsubClientResource.getReady();
         const promises = [];
         for (let i = 0; i < 10; i++) {
@@ -94,8 +96,8 @@ if (process.env.SKIP_CONC) {
 
       it("succeeds exactly once", () => {
         expect(
-            results.filter((result) => responseCodeGroup(result.status) === "2xx")
-                .length
+          results.filter((result) => responseCodeGroup(result.status) === "2xx")
+            .length
         ).toEqual(1);
       });
       it("creates the resource", async () => {
@@ -111,30 +113,30 @@ if (process.env.SKIP_CONC) {
       });
       it("adds the resource in the container listing exactly once", async () => {
         const containerListing = await getContainerMembers(
-            containerUrl,
-            authFetcher
+          containerUrl,
+          authFetcher
         );
-        expect(containerListing.filter((x) => x === resourceUrl).length).toEqual(
-            1
-        );
+        expect(
+          containerListing.filter((x) => x === resourceUrl).length
+        ).toEqual(1);
       });
       ifWps("emits websockets-pubsub on the container exactly once", () => {
         expect(
-            websocketsPubsubClientContainer.received.filter(
-                (x) => x === `pub ${containerUrl}`
-            ).length
+          websocketsPubsubClientContainer.received.filter(
+            (x) => x === `pub ${containerUrl}`
+          ).length
         ).toEqual(1);
       });
       ifWps("emits websockets-pubsub on the resource exactly once", () => {
         expect(
-            websocketsPubsubClientResource.received.filter(
-                (x) => x === `pub ${resourceUrl}`
-            ).length
+          websocketsPubsubClientResource.received.filter(
+            (x) => x === `pub ${resourceUrl}`
+          ).length
         ).toEqual(1);
       });
     });
     describe("Use PATCH 10 times to add triple to the same resource", () => {
-      const {testFolderUrl} = generateTestFolder();
+      const { testFolderUrl } = generateTestFolder();
       let websocketsPubsubClientResource;
       let results;
       let expectedRdf = "";
@@ -153,7 +155,10 @@ if (process.env.SKIP_CONC) {
         });
         await new Promise((resolve) => setTimeout(resolve, waittime));
 
-        websocketsPubsubClientResource = new WPSClient(resourceUrl, authFetcher);
+        websocketsPubsubClientResource = new WPSClient(
+          resourceUrl,
+          authFetcher
+        );
         await websocketsPubsubClientResource.getReady();
         const promises = [];
         for (let i = 0; i < 10; i++) {
@@ -164,9 +169,9 @@ if (process.env.SKIP_CONC) {
               "Content-Type": "text/n3",
             },
             body:
-                "@prefix solid: <http://www.w3.org/ns/solid/terms#>.\n" +
-                "<#patch> a solid:InsertDeletePatch;\n" +
-                `  solid:inserts { ${triple} .}.\n`,
+              "@prefix solid: <http://www.w3.org/ns/solid/terms#>.\n" +
+              "<#patch> a solid:InsertDeletePatch;\n" +
+              `  solid:inserts { ${triple} .}.\n`,
           });
           expectedRdf += `${triple}\n`;
           promises.push(promise);
@@ -182,7 +187,7 @@ if (process.env.SKIP_CONC) {
 
       it("succeeds 10 times", async () => {
         expect(
-            results.filter((x) => responseCodeGroup(x.status) === "2xx").length
+          results.filter((x) => responseCodeGroup(x.status) === "2xx").length
         ).toEqual(10);
       });
 
@@ -196,15 +201,15 @@ if (process.env.SKIP_CONC) {
         rdflib.parse(await result.text(), store2, resourceUrl, "text/turtle");
 
         expect(store2.statements).toEqual(
-            expect.arrayContaining(store1.statements)
+          expect.arrayContaining(store1.statements)
         );
         expect(result.headers.get("Content-Type")).toContain("text/turtle");
       });
       ifWps("emits websockets-pubsub on the resource 10 times", () => {
         expect(
-            websocketsPubsubClientResource.received.filter(
-                (x) => x === `pub ${resourceUrl}`
-            ).length
+          websocketsPubsubClientResource.received.filter(
+            (x) => x === `pub ${resourceUrl}`
+          ).length
         ).toEqual(10);
       });
     });
