+++ test/surface/create-non-container.test.ts
@@ -8,11 +8,11 @@ import { getAuthFetcher } from "solid-auth-fetcher";
 import {
   recursiveDelete,
   getContainerMembers,
-  WPSClient,
+  getStore,
   responseCodeGroup,
   ifWps,
 } from "../helpers/util";
-import { getStore } from "../helpers/util";
+import { WPSClient } from "../helpers/NotificationsClient";
 import { ldp, rdf, space, link } from "rdf-namespaces";
 import * as rdflib from "rdflib";
